new version

+++ lollmsvectordb/vector_database.py
@@ -14,11 +14,12 @@ import hashlib
 from sklearn.neighbors import NearestNeighbors
 from lollmsvectordb.vectorizer import Vectorizer
 from lollmsvectordb.tokenizer import Tokenizer
-from lollmsvectordb.tokenizers.tiktoken_tokenizer import TikTokenTokenizer
+from lollmsvectordb.lollms_tokenizers.tiktoken_tokenizer import TikTokenTokenizer
 from typing import List, Tuple, Optional, Union
-from ascii_colors import ASCIIColors
+from ascii_colors import ASCIIColors, trace_exception
 import pickle
 from pathlib import Path
+from tqdm import tqdm
 
 class VectorDatabase:
     """
@@ -83,11 +84,16 @@ class VectorDatabase:
         self.metrics    = metrics
         self.nn_model = None
         self.vectors = []
-        self.text = []
+        self.texts = []
+        self.documents=[]
+        self.chunks=[]
+        self.titles=[]
+        self.paths=[]
 
-        self._create_tables()
-        self._load_vectors()
-        self.load_first_kneighbors_model()
+        if db_path!="":
+            self._create_tables()
+            self._load_vectors()
+            self.load_first_kneighbors_model()
         self.new_data=False
 
     def _create_tables(self):
@@ -146,12 +152,15 @@ class VectorDatabase:
         return hashlib.sha256(text.encode('utf-8')).hexdigest()
     
     def get_first_vectorizer_name(self) -> Optional[str]:
-        with sqlite3.connect(self.db_path) as conn:
-            cursor = conn.cursor()
-            cursor.execute('SELECT name FROM vectorizer_info ORDER BY id LIMIT 1')
-            result = cursor.fetchone()
-            return result[0] if result else ""
-    
+        if self.db_path!="":
+            with sqlite3.connect(self.db_path) as conn:
+                cursor = conn.cursor()
+                cursor.execute('SELECT name FROM vectorizer_info ORDER BY id LIMIT 1')
+                result = cursor.fetchone()
+                return result[0] if result else ""
+        else:
+            return self.vectorizer.name
+        
     def set_first_vectorizer_name(self, new_name: str) -> bool:
         """
         Sets the name of the first vectorizer in the database.
@@ -162,19 +171,23 @@ class VectorDatabase:
         Returns:
             bool: True if the update was successful, False otherwise.
         """
-        with sqlite3.connect(self.db_path) as conn:
-            cursor = conn.cursor()
-            cursor.execute('SELECT id FROM vectorizer_info ORDER BY id LIMIT 1')
-            result = cursor.fetchone()
-            if result:
-                first_id = result[0]
-                cursor.execute('UPDATE vectorizer_info SET name = ? WHERE id = ?', (new_name, first_id))
-                conn.commit()
-                return True
-            return False
+        if self.db_path!="":
+            with sqlite3.connect(self.db_path) as conn:
+                cursor = conn.cursor()
+                cursor.execute('SELECT id FROM vectorizer_info ORDER BY id LIMIT 1')
+                result = cursor.fetchone()
+                if result:
+                    first_id = result[0]
+                    cursor.execute('UPDATE vectorizer_info SET name = ? WHERE id = ?', (new_name, first_id))
+                    conn.commit()
+                    return True
+                else:
+                    cursor.execute('INSERT INTO vectorizer_info (name) VALUES (?)', (new_name,))
+                return False
+        else:
+            return
 
-
-    def add_document(self, title: str, text: str, path: Union[str, Path]="unknown", chunk_size: int = 512):
+    def add_document(self, title: str, text: str, path: Union[str, Path]="unknown", chunk_size: int = 512, force_update=False):
         """
         Adds a document and its chunks to the database.
 
@@ -190,49 +203,81 @@ class VectorDatabase:
             The size of each chunk (default is 512).
         """
         doc_hash = self._hash_document(text)
-        with sqlite3.connect(self.db_path) as conn:
-            cursor = conn.cursor()
-            cursor.execute('''
-                SELECT id FROM documents WHERE hash = ?
-            ''', (doc_hash,))
-            result = cursor.fetchone()
-            
-            if result is not None:
-                print(f"Document with hash {doc_hash} already exists. It's like déjà vu all over again! 🌀")
-                return
-            
-            cursor.execute('''
-                INSERT INTO documents (hash, title, path) VALUES (?, ?, ?)
-            ''', (doc_hash, title, str(path)))
-            document_id = cursor.lastrowid
 
-            paragraphs = text.split('\n')
-            current_chunk = ""
-            current_tokens = []
+        if self.db_path!="":
+            with sqlite3.connect(self.db_path) as conn:
+                cursor = conn.cursor()
+                cursor.execute('''
+                    SELECT id FROM documents WHERE hash = ?
+                ''', (doc_hash,))
+                result = cursor.fetchone()
+                
+                if result is not None:
+                    if not force_update:
+                        print(f"Document with hash {doc_hash} already exists. It's like déjà vu all over again! 🌀")
+                        return
+                    else:
+                        cursor.execute('''
+                            DELETE FROM documents WHERE hash = ?
+                        ''', (doc_hash,))
+                        conn.commit()
+                
+                cursor.execute('''
+                    INSERT INTO documents (hash, title, path) VALUES (?, ?, ?)
+                ''', (doc_hash, title, str(path)))
+                document_id = cursor.lastrowid
 
-            for paragraph in paragraphs:
-                paragraph_tokens = self.tokenizer.tokenize(paragraph)
-                if len(current_tokens) + len(paragraph_tokens) > chunk_size:
+                paragraphs = text.split('\n')
+                current_chunk = ""
+                current_tokens = []
+                ASCIIColors.multicolor(["Chunking file:",f"{title}"],[ASCIIColors.color_red, ASCIIColors.color_yellow])
+                for paragraph in tqdm(paragraphs):
+                    paragraph_tokens = self.tokenizer.tokenize(paragraph)
+                    if len(current_tokens) + len(paragraph_tokens) > chunk_size:
+                        vector = self.vectorizer.vectorize([current_chunk])[0]
+                        vector_blob = np.array(vector).tobytes()
+                        cursor.execute('''
+                            INSERT INTO chunks (document_id, vector, text, title) VALUES (?, ?, ?, ?)
+                        ''', (document_id, vector_blob, current_chunk, title))
+                        current_chunk = paragraph
+                        current_tokens = paragraph_tokens
+                    else:
+                        current_chunk += "\n" + paragraph
+                        current_tokens.extend(paragraph_tokens)
+
+                if current_chunk:
                     vector = self.vectorizer.vectorize([current_chunk])[0]
                     vector_blob = np.array(vector).tobytes()
                     cursor.execute('''
                         INSERT INTO chunks (document_id, vector, text, title) VALUES (?, ?, ?, ?)
                     ''', (document_id, vector_blob, current_chunk, title))
+
+                conn.commit()
+                self.new_data=True
+        else:
+            paragraphs = text.split('\n')
+            current_chunk = ""
+            current_tokens = []
+            self.documents.append({"title":title,"hash":doc_hash})
+            self.titles.append(title)
+            self.paths.append(path)
+            ASCIIColors.multicolor(["Chunking file:",f"{title}"],[ASCIIColors.color_red, ASCIIColors.color_yellow])
+            for i,paragraph in enumerate(tqdm(paragraphs)):
+                paragraph_tokens = self.tokenizer.tokenize(paragraph)
+                if len(current_tokens) + len(paragraph_tokens) > chunk_size or i==len(paragraphs)-1:
+                    vector = self.vectorizer.vectorize([current_chunk])[0]
+                    vector_blob = np.array(vector).tobytes()
                     current_chunk = paragraph
                     current_tokens = paragraph_tokens
+                    self.chunks.append({
+                        "title":title,
+                        "chunk":current_chunk,
+                        "vector":vector
+                    })
+                    self.texts.append(current_chunk)
                 else:
                     current_chunk += "\n" + paragraph
                     current_tokens.extend(paragraph_tokens)
-
-            if current_chunk:
-                vector = self.vectorizer.vectorize([current_chunk])[0]
-                vector_blob = np.array(vector).tobytes()
-                cursor.execute('''
-                    INSERT INTO chunks (document_id, vector, text, title) VALUES (?, ?, ?, ?)
-                ''', (document_id, vector_blob, current_chunk, title))
-
-            conn.commit()
-            self.new_data=True
 
 
 
@@ -245,17 +290,33 @@ class VectorDatabase:
         doc_hash : str
             The hash of the document to be removed.
         """
-        with sqlite3.connect(self.db_path) as conn:
-            cursor = conn.cursor()
-            cursor.execute('''
-                DELETE FROM chunks WHERE document_id IN (
-                    SELECT id FROM documents WHERE hash = ?
-                )
-            ''', (doc_hash,))
-            cursor.execute('''
-                DELETE FROM documents WHERE hash = ?
-            ''', (doc_hash,))
-            conn.commit()
+        if self.db_path:
+            with sqlite3.connect(self.db_path) as conn:
+                cursor = conn.cursor()
+                # Delete chunks associated with the document
+                cursor.execute('''
+                    DELETE FROM chunks WHERE document_id IN (
+                        SELECT id FROM documents WHERE hash = ?
+                    )
+                ''', (doc_hash,))
+                # Delete the document itself
+                cursor.execute('''
+                    DELETE FROM documents WHERE hash = ?
+                ''', (doc_hash,))
+                conn.commit()
+        else:
+            # Remove document from in-memory storage
+            try:
+                index = [i for i,_ in enumerate(self.chunks) if [doc for doc in self.documents if doc["hash"]==doc_hash][0] == doc_hash]
+                doc_id = self.documents.index(self.chunks[index]["title"])
+                del self.documents[doc_id]
+                del self.titles[doc_id]
+                del self.paths[doc_id]
+                del self.texts[index]
+                del self.chunks[index]
+                print(f"Document with hash '{doc_hash}' removed from in-memory storage.")
+            except:
+                ASCIIColors.error("Document Not found!")
 
     def verify_document(self, text: str) -> bool:
         """
@@ -272,49 +333,68 @@ class VectorDatabase:
             True if the document exists, False otherwise.
         """
         doc_hash = self._hash_document(text)
-        with sqlite3.connect(self.db_path) as conn:
-            cursor = conn.cursor()
-            cursor.execute('''
-                SELECT 1 FROM documents WHERE hash = ?
-            ''', (doc_hash,))
-            return cursor.fetchone() is not None
-
+        if self.db_path!="":
+            with sqlite3.connect(self.db_path) as conn:
+                cursor = conn.cursor()
+                cursor.execute('''
+                    SELECT 1 FROM documents WHERE hash = ?
+                ''', (doc_hash,))
+                return cursor.fetchone() is not None
+        else:
+            return len([d for d in self.documents if d["hash"]==doc_hash])>0
+        
     def _load_vectors(self):
         """
         Loads vectors and their text from the database into memory.
         """
-        with sqlite3.connect(self.db_path) as conn:
-            cursor = conn.cursor()
-            cursor.execute('''
-                SELECT chunks.vector, chunks.text, documents.title 
-                FROM chunks 
-                JOIN documents ON chunks.document_id = documents.id
-            ''')
-            rows = cursor.fetchall()
-            self.vectors = [np.frombuffer(row[0], dtype=np.float32) for row in rows]
-            self.texts = [row[1] for row in rows]
-            self.titles = [row[2] for row in rows]
-
+        if self.db_path!="":
+            with sqlite3.connect(self.db_path) as conn:
+                cursor = conn.cursor()
+                cursor.execute('''
+                    SELECT chunks.vector, chunks.text, documents.title, documents.path 
+                    FROM chunks 
+                    JOIN documents ON chunks.document_id = documents.id
+                ''')
+                rows = cursor.fetchall()
+                self.vectors = [np.frombuffer(row[0], dtype=np.float32) for row in rows]
+                self.texts = [row[1] for row in rows]
+                self.titles = [row[2] for row in rows]
+                self.paths = [row[3] for row in rows]
+        else:
+            ASCIIColors.error("Can't load vectors from database if you don't specify a file path")
     def _update_vectors(self):
         """
         Updates vectors in the database using the current vectorizer.
         """
-        with sqlite3.connect(self.db_path) as conn:
-            cursor = conn.cursor()
-            cursor.execute('SELECT id, text FROM chunks')
-            rows = cursor.fetchall()
-            for row in rows:
-                chunk_id, text = row
-                vector = self.vectorizer.vectorize([text])[0]
-                vector_blob = vector.tobytes()
-                cursor.execute('UPDATE chunks SET vector = ? WHERE id = ?', (vector_blob, chunk_id))
-            conn.commit()
+        if self.db_path!="":
+            with sqlite3.connect(self.db_path) as conn:
+                cursor = conn.cursor()
+                cursor.execute('SELECT id, text FROM chunks')
+                rows = cursor.fetchall()
+                for row in rows:
+                    chunk_id, text = row
+                    vector = self.vectorizer.vectorize([text])[0]
+                    vector_blob = vector.tobytes()
+                    cursor.execute('UPDATE chunks SET vector = ? WHERE id = ?', (vector_blob, chunk_id))
+                conn.commit()
+        else:
+            try:
+                for chunk in self.chunks:
+                    text = chunk["chunk"]
+                    vector = self.vectorizer.vectorize([text])[0]
+                    chunk["vector"] = vector
+                    self.vectors.append(vector)
+            except Exception as ex:
+                trace_exception(ex)
+                ASCIIColors.error("Document Not found!")
 
 
     def store_kneighbors_model(self) -> None:
         """
         Store the KNeighbors model into the SQLite database as the only entry.
         """
+        if self.db_path=="":
+            return
         # Serialize the model using pickle
         model_blob = pickle.dumps(self.n_neighbors)
         
@@ -361,12 +441,21 @@ class VectorDatabase:
         self.n_neighbors = pickle.loads(model_blob[0])
         
 
-    def build_index(self):
+    def build_index(self, force_reindex= False):
         """
         Builds the nearest neighbors index using the loaded vectors.
         """
-        if self.vectorizer.name != self.get_first_vectorizer_name():
-            ASCIIColors.warning("Detected a change in the vectorizer. Revectorizing the whole database")
+        if self.db_path=="":
+            self._update_vectors()
+            self.nn_model = NearestNeighbors(n_neighbors=self.n_neighbors, algorithm=self.algorithm, metric='cosine')
+            if len(self.vectors)>0:
+                self.nn_model.fit(self.vectors)
+                self.store_kneighbors_model()
+            return
+        
+        if self.vectorizer.name != self.get_first_vectorizer_name() or force_reindex:
+            if self.vectorizer.name != self.get_first_vectorizer_name() and self.get_first_vectorizer_name()!="":
+                ASCIIColors.warning("Detected a change in the vectorizer. Revectorizing the whole database")
             self._update_vectors()
             self.set_first_vectorizer_name(self.vectorizer.name)
         else:
@@ -390,7 +479,7 @@ class VectorDatabase:
         Returns:
         --------
         list of tuples
-            A list of tuples containing the vector, text, and distance of the nearest neighbors.
+            A list of tuples containing the vector, text, title, path, and distance of the nearest neighbors.
         """
         if self.nn_model is None:
             raise ValueError("Index not built. Call build_index() before searching.")
@@ -400,7 +489,7 @@ class VectorDatabase:
         
         query_vector = self.vectorizer.vectorize([query_data])[0]
         distances, indices = self.nn_model.kneighbors([query_vector], n_neighbors=n_results)
-        results = [(self.vectors[idx], self.texts[idx], self.titles[idx], distances[0][i]) for i, idx in enumerate(indices[0])]
+        results = [(self.vectors[idx], self.texts[idx], self.titles[idx], self.paths[idx], distances[0][i]) for i, idx in enumerate(indices[0])]
         return results
 
     def save(self):
@@ -436,11 +525,11 @@ class VectorDatabase:
 if __name__ == "__main__":
     # Example with TFIDFVectorizer
     from lollmsvectordb import TFIDFVectorizer
-    from lollmsvectordb.vectorizers.bert_vectorizer import BERTVectorizer
+    from lollmsvectordb.lollms_vectorizers.bert_vectorizer import BERTVectorizer
     tfidf_vectorizer = BERTVectorizer()
     tfidf_vectorizer.fit(["This is a sample text.", "Another sample text.", "Another sample text 1.", "Another sample text 2.", "Another sample text 3.", "Another sample text 4."])
     tokenizer = TikTokenTokenizer()
-    db = VectorDatabase("vector_db.sqlite", tfidf_vectorizer, tokenizer)
+    db = VectorDatabase("vector_db.sqlite", tfidf_vectorizer, tokenizer) # 
 
     # Add multiple documents to the database
     documents = [
@@ -463,5 +552,5 @@ if __name__ == "__main__":
     results = db.search(query, n_results=3)
 
     # Print the search results
-    for vector, texts, titles, distance in results:
+    for vector, text, title, distance in results:
         print(f"Title: {title}, Text: {text}, Distance: {distance}")
